split files (pr comment)
