+++ core/runtime/common/runtime_context.cpp
@@ -9,7 +9,6 @@
 #include "runtime/runtime_context.hpp"
 
 #include "blockchain/block_header_repository.hpp"
-#include "runtime/common/stack_limiter.hpp"
 #include "runtime/common/uncompress_code_if_needed.hpp"
 #include "runtime/instance_environment.hpp"
 #include "runtime/memory_provider.hpp"
@@ -18,6 +17,7 @@
 #include "runtime/module_instance.hpp"
 #include "runtime/module_repository.hpp"
 #include "runtime/trie_storage_provider.hpp"
+#include "runtime/wabt/instrument.hpp"
 #include "storage/trie/polkadot_trie/trie_error.hpp"
 
 OUTCOME_CPP_DEFINE_CATEGORY(kagome::runtime, Error, e) {
