+++ core/runtime/wabt/stack_limiter.cpp
@@ -4,56 +4,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "runtime/common/stack_limiter.hpp"
+#include "runtime/wabt/stack_limiter.hpp"
 
 #include "common/visitor.hpp"
 #include "log/logger.hpp"
 #include "log/profiling_logger.hpp"
-
-#include <wabt/binary-reader-ir.h>
-#include <wabt/binary-reader.h>
-#include <wabt/binary-writer.h>
-#include <wabt/error-formatter.h>
-#include <wabt/ir.h>
-#include <wabt/validator.h>
+#include "runtime/wabt/util.hpp"
 
 namespace kagome::runtime {
-  WabtOutcome<void> wabtTry(auto &&f) {
-    wabt::Errors errors;
-    if (wabt::Failed(f(errors))) {
-      return StackLimiterError{
-          wabt::FormatErrorsToString(errors, wabt::Location::Type::Binary)};
-    }
-    return outcome::success();
-  }
-
-  WabtOutcome<void> wabtDecode(wabt::Module &module, common::BufferView code) {
-    return wabtTry([&](wabt::Errors &errors) {
-      return wabt::ReadBinaryIr(
-          "",
-          code.data(),
-          code.size(),
-          wabt::ReadBinaryOptions({}, nullptr, true, false, false),
-          &errors,
-          &module);
-    });
-  }
-
-  WabtOutcome<void> wabtValidate(const wabt::Module &module) {
-    return wabtTry([&](wabt::Errors &errors) {
-      return wabt::ValidateModule(&module, &errors, wabt::ValidateOptions{});
-    });
-  }
-
-  WabtOutcome<common::Buffer> wabtEncode(const wabt::Module &module) {
-    wabt::MemoryStream s;
-    if (wabt::Failed(wabt::WriteBinaryModule(
-            &s, &module, wabt::WriteBinaryOptions({}, false, false, true)))) {
-      return StackLimiterError{"Failed to serialize WASM module"};
-    }
-    return common::Buffer{std::move(s.output_buffer().data)};
-  }
-
   namespace detail {
 
     template <bool IsConst, typename T>
@@ -155,9 +113,9 @@ namespace kagome::runtime {
             frames_{},
             logger_{std::move(logger)} {}
 
-      outcome::result<void, StackLimiterError> unreachable() {
+      WabtOutcome<void> unreachable() {
         if (frames_.empty()) {
-          return StackLimiterError{"Stack must not be empty"};
+          return WabtError{"Stack must not be empty"};
         }
         frames_.back().is_polymorphic = true;
         return outcome::success();
@@ -184,8 +142,8 @@ namespace kagome::runtime {
         });
       }
 
-      outcome::result<void, StackLimiterError> push_frame(
-          MaybeConst<wabt::IfExpr &> branch, bool check_frame_boundary) {
+      WabtOutcome<void> push_frame(MaybeConst<wabt::IfExpr &> branch,
+                                   bool check_frame_boundary) {
         uint32_t end_arity = branch.true_.decl.GetNumResults() != 0;
         uint32_t branch_arity = end_arity;
         auto res = pop_values(1);
@@ -214,9 +172,9 @@ namespace kagome::runtime {
         });
       }
 
-      outcome::result<void, StackLimiterError> pop_frame() {
+      WabtOutcome<void> pop_frame() {
         if (frames_.empty()) {
-          return StackLimiterError{"Stack is empty"};
+          return WabtError{"Stack is empty"};
         }
         height_ = frames_.back().start_height;
         push_values(frames_.back().end_value_num);
@@ -238,24 +196,23 @@ namespace kagome::runtime {
         SL_TRACE(logger_, "push {}, now height_ {}", num, height_);
       }
 
-      outcome::result<void, StackLimiterError> pop_values(uint32_t num) {
+      WabtOutcome<void> pop_values(uint32_t num) {
         if (num == 0) {
           return outcome::success();
         }
         if (frames_.empty()) {
-          return StackLimiterError{"Stack is empty"};
+          return WabtError{"Stack is empty"};
         }
         SL_TRACE(logger_, "pop {}, now height_ {}", num, height_ - num);
         if (height_ - num < frames_.back().start_height) {
           if (!frames_.back().is_polymorphic) {
-            return StackLimiterError{
-                "Popping values not pushed in the current frame"};
+            return WabtError{"Popping values not pushed in the current frame"};
           } else {
             return outcome::success();
           }
         }
         if (height_ < num) {
-          return StackLimiterError{"Stack underflow"};
+          return WabtError{"Stack underflow"};
         }
         height_ -= num;
         return outcome::success();
@@ -265,7 +222,7 @@ namespace kagome::runtime {
         return frames_.empty();
       }
 
-      outcome::result<void, StackLimiterError> advance() {
+      WabtOutcome<void> advance() {
         bool is_over = false;
         do {
           auto &frame = frames_.back();
@@ -309,11 +266,10 @@ namespace kagome::runtime {
         return &frames_.back();
       }
 
-      [[nodiscard]] outcome::result<std::reference_wrapper<const StackFrame>,
-                                    StackLimiterError>
+      [[nodiscard]] WabtOutcome<std::reference_wrapper<const StackFrame>>
       get_frame(size_t idx_from_top) const {
         if (frames_.size() <= idx_from_top) {
-          return StackLimiterError{"Stack frame underflow"};
+          return WabtError{"Stack frame underflow"};
         }
         return frames_.at(frames_.size() - idx_from_top - 1);
       }
@@ -327,10 +283,9 @@ namespace kagome::runtime {
     using ConstStack = Stack<true>;
     using MutStack = Stack<false>;
 
-    outcome::result<uint32_t, StackLimiterError> compute_stack_cost(
-        const log::Logger &logger,
-        const wabt::Func &func,
-        const wabt::Module &module) {
+    WabtOutcome<uint32_t> compute_stack_cost(const log::Logger &logger,
+                                             const wabt::Func &func,
+                                             const wabt::Module &module) {
       uint32_t locals_num = func.GetNumLocals();
 
       ConstStack stack{logger};
@@ -410,7 +365,7 @@ namespace kagome::runtime {
               OUTCOME_TRY(frame, stack.get_frame(v.index()));
               uint32_t arity = frame.get().branch_value_num;
               if (arity != target_arity) {
-                return StackLimiterError{
+                return WabtError{
                     "All jump-targets should have equal frame arities"};
               }
             }
@@ -527,9 +482,8 @@ namespace kagome::runtime {
             break;
           }
           default:
-            return StackLimiterError{
-                fmt::format("Unsupported instruction: {}",
-                            wabt::GetExprTypeName(expr.type()))};
+            return WabtError{fmt::format("Unsupported instruction: {}",
+                                         wabt::GetExprTypeName(expr.type()))};
         }
         if (!pushed_frame) {
           if (auto expr_opt = stack.advance(); !expr_opt.has_value()) {
@@ -591,7 +545,7 @@ namespace kagome::runtime {
       return next_it;
     }
 
-    outcome::result<void, StackLimiterError> instrument_func(
+    WabtOutcome<void> instrument_func(
         wabt::Func &func,
         const wabt::Var &stack_height,
         uint32_t stack_limit,
@@ -664,7 +618,7 @@ namespace kagome::runtime {
       return outcome::success();
     }
 
-    outcome::result<void, StackLimiterError> generate_thunks(
+    WabtOutcome<void> generate_thunks(
         const log::Logger &logger,
         wabt::Module &module,
         const wabt::Var &stack_height,
@@ -702,7 +656,7 @@ namespace kagome::runtime {
               break;
             }
             default:
-              return StackLimiterError{
+              return WabtError{
                   fmt::format("Unsupported element expression of type {}",
                               GetExprTypeName(expr.type()))};
           }
@@ -766,7 +720,7 @@ namespace kagome::runtime {
               break;
             }
             default:
-              return StackLimiterError{
+              return WabtError{
                   fmt::format("Invalid element expression of type {}",
                               GetExprTypeName(expr.type()))};
           }
@@ -834,103 +788,12 @@ namespace kagome::runtime {
       common::BufferView uncompressed_wasm, size_t stack_limit) {
     auto logger = stackLimiterLog();
     KAGOME_PROFILE_START_L(logger, read_ir);
-    wabt::Module module;
-    OUTCOME_TRY(wabtDecode(module, uncompressed_wasm));
+    OUTCOME_TRY(module, wabtDecode(uncompressed_wasm));
     KAGOME_PROFILE_END_L(logger, read_ir);
 
     OUTCOME_TRY(instrumentWithStackLimiter(module, stack_limit));
 
     KAGOME_PROFILE_START_L(logger, serialize_wasm);
     return wabtEncode(module);
-  }
-
-  WabtOutcome<void> convertMemoryImportIntoExport(wabt::Module &module) {
-    for (auto it = module.fields.begin(); it != module.fields.end(); ++it) {
-      auto import = dynamic_cast<const wabt::ImportModuleField *>(&*it);
-      if (not import) {
-        continue;
-      }
-      auto memory = dynamic_cast<wabt::MemoryImport *>(import->import.get());
-      if (not memory) {
-        continue;
-      }
-      if (std::any_of(module.fields.begin(),
-                      module.fields.end(),
-                      [&](const wabt::ModuleField &field) {
-                        return field.type() == wabt::ModuleFieldType::Memory;
-                      })) {
-        return StackLimiterError{"unexpected MemoryModuleField"};
-      }
-      auto import_it = std::find(
-          module.imports.begin(), module.imports.end(), import->import.get());
-      if (import_it == module.imports.end()) {
-        return StackLimiterError{"inconsistent Module.imports"};
-      }
-      auto memory_it = std::find(
-          module.memories.begin(), module.memories.end(), &memory->memory);
-      if (memory_it == module.memories.end()) {
-        return StackLimiterError{"inconsistent Module.memories"};
-      }
-      auto memory2 = std::make_unique<wabt::MemoryModuleField>();
-      memory2->memory.page_limits = memory->memory.page_limits;
-      auto export_ = std::make_unique<wabt::ExportModuleField>();
-      export_->export_ = {
-          import->import->field_name,
-          wabt::ExternalKind::Memory,
-          wabt::Var{0, {}},
-      };
-      module.imports.erase(import_it);
-      module.memories.erase(memory_it);
-      module.fields.erase(it);
-      --module.num_memory_imports;
-      module.AppendField(std::move(memory2));
-      module.AppendField(std::move(export_));
-      break;
-    }
-    return outcome::success();
-  }
-
-  WabtOutcome<void> setupMemoryAccordingToHeapAllocStrategy(
-      wabt::Module &module, const HeapAllocStrategy &config) {
-    for (auto it = module.fields.begin(); it != module.fields.end(); ++it) {
-      auto memory = dynamic_cast<wabt::MemoryModuleField *>(&*it);
-      if (not memory) {
-        continue;
-      }
-      auto &limit = memory->memory.page_limits;
-      auto init = limit.initial;
-      if (auto v = boost::get<HeapAllocStrategyDynamic>(&config)) {
-        if (auto &max = v->maximum_pages) {
-          limit = {std::min<uint64_t>(init, *max), *max};
-        } else {
-          limit = wabt::Limits{init};
-        }
-      } else {
-        auto max =
-            init + boost::get<HeapAllocStrategyStatic>(config).extra_pages;
-        limit = {max, max};
-      }
-    }
-    return outcome::success();
-  }
-
-  WabtOutcome<common::Buffer> prepareBlobForCompilation(
-      common::BufferView code, const MemoryLimits &config) {
-    wabt::Module module;
-    OUTCOME_TRY(wabtDecode(module, code));
-    if (config.max_stack_values_num) {
-      OUTCOME_TRY(
-          instrumentWithStackLimiter(module, *config.max_stack_values_num));
-    }
-    OUTCOME_TRY(convertMemoryImportIntoExport(module));
-    OUTCOME_TRY(setupMemoryAccordingToHeapAllocStrategy(
-        module, config.heap_alloc_strategy));
-    OUTCOME_TRY(wabtValidate(module));
-    return wabtEncode(module);
-  }
-
-  WabtOutcome<common::Buffer> InstrumentWasm::instrument(
-      common::BufferView code, const MemoryLimits &config) const {
-    return prepareBlobForCompilation(code, config);
   }
 }  // namespace kagome::runtime
