+++ test/core/runtime/runtime_test_base.hpp
@@ -45,6 +45,7 @@
 #include "runtime/executor.hpp"
 #include "runtime/module.hpp"
 #include "runtime/runtime_context.hpp"
+#include "runtime/wabt/instrument.hpp"
 #include "storage/in_memory/in_memory_storage.hpp"
 #include "testutil/literals.hpp"
 #include "testutil/outcome.hpp"
